% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addextradatevariables.R
\name{addExtraDateVariables}
\alias{addExtraDateVariables}
\title{Add extra date variables to a data frame}
\usage{
addExtraDateVariables(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame to which the date variables are 
appended.}
}
\value{
Returns a data frame.
}
\description{
\code{addExtraDateVariables} appends extra date variables to a data frame.
}
\details{
These variables are integers for convenience use in filtering:

\itemize{
  \item \code{Year}: The year
  \item \code{Month}: The month
  \item \code{DayOfMonth}: The date (1-31)
  \item \code{DayOfYear}: The so-called \dQuote{julian day} as an integer 
    (1-366)
}
}
\examples{
\dontrun{addExtraDateVariables(bothStations)}
}

