% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\name{allData}
\alias{allData}
\title{Returns data frame with all weather observations}
\usage{
allData()
}
\value{
Returns a data frame.
}
\description{
\code{allData} takes \code{\link{bothStations}} and 
\code{\link{updatedData}} and combines them.
}
\details{
This is a convenience wrapper that uses \code{\link{combineDataFrames}} to 
combine the two data frames, sorted by date.
}
\examples{
allData()
}

