% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendcsv.R
\name{append.csv}
\alias{append.csv}
\title{Appending a row to a CSV.}
\usage{
append.csv(appendData, CsvFile)
}
\arguments{
\item{appendData}{The data frame with the information for the row(s) to be 
appended to the CSV.}

\item{CsvFile}{The file name of the CSV to which \code{data} is appended.}
}
\value{
No return; appends to the specified file.
}
\description{
\code{append.csv} appends a row or rows to an existing CSV file.
}
\details{
This is a convenience wrapper that uses \code{write.table} to append to 
a CSV.

This function is still not well-developed. Users must ensure that the
data to be appended to the CSV file has the same fields in the same order.
No checking is performed by \code{append.csv}
}
\examples{
\dontrun{append.csv()}
}

