% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkyear.R
\name{checkYear}
\alias{checkYear}
\title{Check whether or not the provided year is valid.}
\usage{
checkYear(uncheckedYear)
}
\arguments{
\item{uncheckedYear}{A value for the year in question.}
}
\value{
Returns a string with the year.
}
\description{
\code{checkYear} returns the year provided to it, unless it is invalid (in
which case, it returns the current year).
}
\details{
This function determines that the year in question is:
\itemize{
  \item not NULL
  \item not earlier than 1724 (the year the Fahrenheit scale was developed)
  \item not later than the current year
}
}
\examples{
\dontrun{checkYear("2016")}  # "2016"
\dontrun{checkYear("1776")}  # "2016" (< 1724 becomes current year)
\dontrun{checkYear("1984")}  # "1984"
\dontrun{checkYear("2020")}  # "2016" (future year becomes current year)
}

