% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedataframes.R
\name{combineDataFrames}
\alias{combineDataFrames}
\title{Combines two data frames with weather observations}
\usage{
combineDataFrames(dfOne, dfTwo)
}
\arguments{
\item{dfOne}{The first data frame.}

\item{dfTwo}{The second data frame, which is added to the first.}
}
\value{
Returns a data frame.
}
\description{
\code{combineDataFrames} takes two data frames and combines them.
}
\details{
This is a convenience wrapper that uses \code{rbind} to append to 
a CSV, then sorts the data frame by date.

No error checking is done: Users should ensure that both data frames are
in the proper format.
}
\examples{
combineDataFrames(earlyDowntownData, airportData)
}

