% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeconsecutiveprecipvariables.R
\name{computeConsecutivePrecipVariables}
\alias{computeConsecutivePrecipVariables}
\title{Calculate consecutive days of precipitation}
\usage{
computeConsecutivePrecipVariables(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame to which the
\code{consecutivePrecipitation} and \code{consecutiveSnowfall} variables
will be appended.}
}
\value{
Returns a data frame.
}
\description{
\code{computeConsecutivePrecipVariables} appends
\code{consecutivePrecipitation} and \code{consecutiveSnowfall} variables,
each containing an integer representing the number of consecutive days with
recorded precipitation and snowfall, respectively.
}
\details{
This function sorts \code{originalFrame} by date, and then the two
variables (\code{consecutivePrecipitation} and \code{consecutiveSnowfall})
are initialized with a value of 0. Then, a for loop is used which leaves
that initialized value alone if \code{WithPrecipitation} (or
\code{WithSnowfall}) is \code{FALSE} or \code{NA}. If the
\dQuote{With* variable} is \code{TRUE}, the value of the previous day's
\dQuote{consecutive* variable} is taken and incremented by 1.

This function assumes that \code{originalFrame} contains the precipitation
variables (which are usually created by
\code{\link[=convertCsvToNumericAndLogical]{convertCsvToNumericAndLogical()}}) and assumes that
\code{originalFrame} is sorted from earliest to most recent. If either of
these is not the case, this function will fail or return data which are
incorrect.

This function has been extended to also create
\code{consecutiveMeasurablePrecipitation} (counting days with precipitation ≥
0.01").
}
\examples{
computeConsecutivePrecipVariables(convertCsvToNumericAndLogical(
                                    airportData))
}
