% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computecumulativeprecipitation.R
\name{computeCumulativePrecipitation}
\alias{computeCumulativePrecipitation}
\title{Compute cumulative precipitation variables}
\usage{
computeCumulativePrecipitation(originalFrame = orfwx::allData(),
  daysWeather = NULL)
}
\arguments{
\item{originalFrame}{(optional) The data frame to which the \code{MTDPrecip}
and \code{YTDPrecip} variables are appended. Defaults to \code{allData()}.}

\item{daysWeather}{(optional) The weather for a date not yet included in
the \code{originalFrame} data frame, usually passed by the
\code{singleDaysWeather} function.}
}
\value{
Returns a data frame.
}
\description{
\code{computeCumulativePrecipitation} returns a data frame with
month-to-date and year-to-date variables for precipitation.
}
\details{
Details section to be written
}
\examples{
\dontrun{computeCumulativePrecipitation()}
}
