% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computecumulativeprecipitationrecords.R
\name{computeCumulativePrecipitationRecords}
\alias{computeCumulativePrecipitationRecords}
\title{Compute cumulative precipitation variables}
\usage{

  computeCumulativePrecipitationRecords(originalFrame = orfwx::computeCumulativePrecipitation(),
  ccprMonth = format(orfwx::yesterdate(), "\%m"))
}
\arguments{
\item{originalFrame}{(optional) The data frame to which the \code{MTDPrecip} and 
\code{YTDPrecip} variables are appended. Defaults to \code{allData()}.}
}
\value{
Returns a data frame.
}
\description{
\code{computeCumulativePrecipitationRecords} returns a data frame with
the maximum and minimum values of \code{MTDPrecip}.
}
\details{
Details section to be written
}
\examples{
\dontrun{computeCumulativePrecipitation()}
}

