% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computecumulativepreciprecords.R
\name{computeCumulativePrecipRecords}
\alias{computeCumulativePrecipRecords}
\title{Compute cumulative precipitation variables}
\usage{

  computeCumulativePrecipRecords(originalFrame = orfwx::computeCumulativePrecipitation(),
  ccprMonth = format(orfwx::yesterdate(), "\%m"), showYear = FALSE,
  includeNormals = FALSE,
  showLeapDay = orfwx::is.leapYear(as.integer(format(orfwx::yesterdate(),
  "\%Y"))))
}
\arguments{
\item{originalFrame}{(optional) The data frame to which the \code{MTDPrecip} 
and \code{YTDPrecip} variables are appended. Defaults to \code{allData()}.}

\item{ccprMonth}{(optional) The month for which precipitation records are
desired. Defaults to the current month (except on the 1st of the month,
when it defaults to the previous month).}

\item{showYear}{(optional) Whether or not to show the current or most recent
year's data for comparison (defaults to \code{FALSE}).}

\item{includeNormals}{(optional) Whether or not to show the 1981-2010 
normals for comparison (defaults to \code{FALSE}).}

\item{showLeapDay}{(optional) Whether or not to show data for February 29th
(defaults to \code{FALSE}, unless the current year is a leap year).}
}
\value{
Returns a data frame.
}
\description{
\code{computeCumulativePrecipitationRecords} returns a data frame with
the maximum and minimum values of \code{MTDPrecip}.
}
\details{
Details section to be written
}
\examples{
\dontrun{computeCumulativePrecipitation()}
}

