% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computetemperaturespread.R
\name{computeTemperatureSpread}
\alias{computeTemperatureSpread}
\title{Add variable for temperature spread to a data frame}
\usage{
computeTemperatureSpread(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame to which the \code{TemperatureSpread}
variable is appended.}
}
\value{
Returns a data frame.
}
\description{
\code{computeTemperatureSpread} appends a variable to the data frame with
the difference between the day's high and low temperatures.
}
\details{
The difference between \code{MaxTemperature} and \code{MinTemperature} is
computed and added to a \code{TemperatureSpread} variable in the data frame.

This will stop with an error if those columns are not present in
\code{originalFrame}.
}
\examples{
computeTemperatureSpread(airportData)
}
