% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertcsvtonumericandlogical.R
\name{convertCsvToNumericAndLogical}
\alias{convertCsvToNumericAndLogical}
\title{Convert CSV values for precipitation and snowfall to numeric and logical}
\usage{
convertCsvToNumericAndLogical(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame to which the date variables are 
appended.}
}
\value{
Returns a data frame.
}
\description{
\code{convertCsvToNumericAndLogical} adds extra variables to express
\code{CsvPrecipitation} and \code{CsvSnowfall} in numeric and logical
variables.
}
\details{
The variables added are:

\itemize{
  \item \code{PrecipitationInches}: The precipitation, accurate to the
    hundredth of an inch (numeric)
  \item \code{WithPrecipitation}: Logical value indicating whether or not
    precipitation fell on this day
  \item \code{SnowfallInches}: The snowfall, accurate to the tenth of an
    inch (numeric)
  \item \code{WithSnowfall}: Logical value indicating whether or not snow
    fell on this day
}
}
\examples{
\dontrun{convertCsvToNumericAndLogical(bothStations)}
}

