% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensuredirectoryexists.R
\name{ensureDirectoryExists}
\alias{ensureDirectoryExists}
\title{Ensures directory exists, optionally create it}
\usage{
ensureDirectoryExists(whichDirectory = "~/.orfwx", createIfMissing = FALSE)
}
\arguments{
\item{whichDirectory}{The directory that is checked for existence. Defaults
to \code{~/.orfwx}}

\item{createIfMissing}{If \code{TRUE}, create the directory if it does not
exist.}
}
\value{
Logical value: \code{TRUE} if the directory exists (either it
  already existed or was created if \code{createIfMissing == TRUE}) and
  \code{FALSE} if the directory does not already exist.
}
\description{
\code{ensureDirectoryExists} ensures that the a directory exists, and, if 
it does not, can create it.
}
\details{
By default, this checks for the presence of \code{~/.orfwx} but does not
create that directory if it is missing. Both of those options can be
overridden by parameters.
}
\examples{
\dontrun{
ensureDirectoryExists("~/.orfwx")
}
}

