% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmostrecentdate.R
\name{findMostRecentDate}
\alias{findMostRecentDate}
\title{Find most recent date in data}
\usage{
findMostRecentDate(fmrdSource = orfwx::updatedData())
}
\arguments{
\item{fmrdSource}{The data frame to check. (Defaults to `updatedData()`).}
}
\value{
Returns a date.
}
\description{
\code{findMostRecentDate} checks a data frame for the most recent date for
which observations are included.
}
\details{
Details section to be written.
}
\examples{
\dontrun{
findMostRecentDate()
}
}

