% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstsunday.R
\name{firstSunday}
\alias{firstSunday}
\title{Find the first Sunday in a month.}
\usage{
firstSunday(fsMonth, fsYear)
}
\arguments{
\item{fsMonth}{An integer representing the month of the year (1=January,
12=December) for which the first Sunday is desired.}

\item{fsYear}{An integer representing the year of the month for which the
first Sunday of the month is desired.}
}
\value{
Integer
}
\description{
\code{firstSunday} finds the date of the first Sunday in a month.
}
\details{
This function takes the provided month and year and returns the date of the
first Sunday of the month. For example, January 2017 begins with a Sunday, so
\code{firstSunday(1, 2017)} returns 1. But February 2017 begins with a
Wednesday, so \code{firstSunday(2, 2017)} returns 5, because February 5th is
the first Sunday of the month.
}
\examples{
firstSunday(1, 2017) # Returns 1
firstSunday(2, 2017) # Returns 5
}
