% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getupdatedcsv.R
\name{getUpdatedCsv}
\alias{getUpdatedCsv}
\title{Combines two data frames with weather observations}
\usage{
getUpdatedCsv()
}
\value{
Downloads a file to the working directory.
}
\description{
\code{getUpdatedCsv} takes two data frames and combines them.
}
\details{
This is a convenience wrapper that uses \code{httr} to download an
\code{updates.csv} file
}
\examples{
\dontrun{getUpdatedCsv()}
}

