% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-leapyear.R
\name{is.leapYear}
\alias{is.leapYear}
\title{Determining whether or not this is a leap year.}
\usage{
is.leapYear(year)
}
\arguments{
\item{year}{An integer value for the year in question.}
}
\value{
Returns a logical value (\code{TRUE} or \code{FALSE}).
}
\description{
\code{is.leapYear} returns \code{TRUE} or \code{FALSE}, depending on
whether or not the year passed to it is a leap year or not.
}
\details{
This function determines that the year in question is:
\itemize{
  \item evenly divisible by 4, and
  \item not evenly divisible by 100, or
  \item \emph{(ignoring the previous conditions)} evenly divisible by 400
}
}
\examples{
is.leapYear(2016)  # TRUE (divisible by 4)
is.leapYear(2015)  # FALSE (not divisible by 4)
is.leapYear(1900)  # FALSE (divisible by 100)
is.leapYear(2000)  # TRUE (divisible by 400)
}

