% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalsuffix.R
\name{ordinalSuffix}
\alias{ordinalSuffix}
\title{Add an ordinal suffix to an integer}
\usage{
ordinalSuffix(osInteger)
}
\arguments{
\item{osInteger}{The number to convert.}
}
\value{
Returns a string.
}
\description{
\code{ordinalSuffix} returns the integer provided to it as a string, with the
ordinal suffix (\emph{st}, \emph{nd}, \emph{rd}, or \emph{th}) appended.
}
\details{
This function determines that the year in question is:
\itemize{
\item not NULL
\item not earlier than 1724 (the year the Fahrenheit scale was developed)
\item not later than the current year
}
}
\examples{
ordinalSuffix(1)  # Returns "1st"
ordinalSuffix(2)  # Returns "2nd"
ordinalSuffix(3)  # Returns "3rd"
ordinalSuffix(4)  # Returns "4th"
ordinalSuffix(11)  # Returns "11th"
}
