% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padsingledigitinteger.R
\name{padSingleDigitInteger}
\alias{padSingleDigitInteger}
\title{Pad a number with a zero to create a two-character string (if single digit).}
\usage{
padSingleDigitInteger(theInteger)
}
\arguments{
\item{theInteger}{Any integer.}
}
\value{
Intended to return a two-character string
}
\description{
\code{padSingleDigitInteger} returns a two-character string, padding a
single-digit integer with a leading zero, if required.
}
\details{
This function currently converts \code{theInteger} into an integer and
returns that integer if greater than 10; otherwise, it adds a leading
zero, intended to create a two-character string.

The input is not currently checked for all necessary correctness.
For sensible results, \code{TheInteger} should be a positive integer,
less than or equal to 99.
}
\examples{
padSingleDigitInteger(20)  # 20
padSingleDigitInteger(9)  # "09"
padSingleDigitInteger(0)  # "00"
}

