% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmintempoverhistory.R
\name{plotMinTempOverHistory}
\alias{plotMinTempOverHistory}
\title{Plot historical low tempertures}
\usage{
plotMinTempOverHistory(plotDate = searchDate(), daysWeather = NULL,
  twoTicks = TRUE, fiveTicks = FALSE, tenTicks = TRUE)
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}

\item{daysWeather}{(optional) The weather for a date not yet included in
the \code{mutatedBothStations} dataset, usually passed by the
\code{singleDaysWeather} function.}

\item{twoTicks}{(optional) Writes half ticks (defaults to \code{TRUE}).}

\item{fiveTicks}{(optional) Writes fifth ticks (defaults to \code{FALSE}).}

\item{tenTicks}{(optional) Writes tenth ticks (defaults to \code{TRUE}).}
}
\value{
Returns a barplot.
}
\description{
\code{plotMinTempOverHistory} returns a barplot with the low temperatures
of one day each year.
}
\details{
If \code{daysWeather} is passed (usually using 
\code{\link{singleDaysWeather}}), the function ensures that its weather is 
for the same date as that provided by \code{plotDate} (either passed to the 
function or the default value of the current system date). If they do not 
match, the function terminates with an error message.

The data from \code{daysWeather} is added to that provided by 
\code{\link{mutatedBothStations}} and the weather on the same date each year
is compared.

A barplot is plotted by \code{\link{plotWithManyBars}} with the low 
temperatures for all days. The bar for the year provided by 
\code{daysWeather} (or, if \code{daysWeather} is not provided, the current 
year) is highlighted with a bar of a different color in the barplot.
}
\examples{
plotMinTempOverHistory(searchDate(11, 26))  # plot for November 26th
}

