% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmtdprecipitation.R
\name{plotMTDPrecipitation}
\alias{plotMTDPrecipitation}
\title{Plot month-to-date precipitation}
\usage{
plotMTDPrecipitation()
}
\value{
Returns a plot.
}
\description{
\code{plotMTDPrecipitation} creates a barplot with the precipitation
of the month to date.
}
\details{
Three lines are displayed: a red line with the highest month-to-date
precipitation recorded by that point in the month, a blue line with the
lowest month-to-date precipitation recorded by that point in the month, and
a black line with the curent year's month-to-date precipitation.
}
\examples{
plotMTDPrecipitation()
}

