% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotprecipitationoverhistory.R
\name{plotPrecipitationOverHistory}
\alias{plotPrecipitationOverHistory}
\title{Plot historical daily precipitation}
\usage{
plotPrecipitationOverHistory(plotDate = searchDate())
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}
}
\value{
Returns a barplot.
}
\description{
\code{plotPrecipitationOverHistory} returns a barplot with the precipitation
of one day each year.
}
\details{
This code makes my head hurt, because it was some of the earliest R code I
wrote, when I was mostly in make-it-work mode, with huge gaps in my
understanding of R (and of barplot particularly)
}
\examples{
plotPrecipitationOverHistory(searchDate(11, 26))  # plot for November 26th
}

