% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwarmestmaxtempoverhistory.R
\name{plotWarmestMaxTempOverHistory}
\alias{plotWarmestMaxTempOverHistory}
\title{Plot warmest historical high tempertures on a single date}
\usage{
plotWarmestMaxTempOverHistory(plotDate = searchDate())
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}
}
\value{
Returns a barplot.
}
\description{
\code{plotWarmestMaxTempOverHistory} returns a barplot with the warmest 
high temperatures of one day of the year.
}
\details{
This code makes my head hurt, because it was some of the earliest R code I
wrote, when I was mostly in make-it-work mode, with huge gaps in my
understanding of R (and of barplot particularly)
}
\examples{
plotWarmestMaxTempOverHistory(searchDate(11, 26))  # plot for November 26th
}

