% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwithmanybars.R
\name{plotWithManyBars}
\alias{plotWithManyBars}
\title{Create a barplot with temperature or precipitation data}
\usage{
plotWithManyBars(sortedData, sortedDataFrame, titlePlotWmb, yAxisLabelPlotWmb,
  plottingPrecip = FALSE, showAllLabels = FALSE,
  highlightYear = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{sortedData}{A vector used for the values to be shown on the barplot.}

\item{sortedDataFrame}{A data frame, from which \code{sortedData} has
usually been excerpted.}

\item{titlePlotWmb}{The title of the barplot}

\item{yAxisLabelPlotWmb}{The label for the y axis}

\item{plottingPrecip}{A logical value indicating if we are dealing with
precipitation}

\item{showAllLabels}{A logical value indicating if we are showing all
labels (if \code{FALSE}, labels will only be shown on the max and min
values and on the current year)}

\item{highlightYear}{The year to highlight in the graph (defaults to the
current year)}
}
\value{
Returns a barplot.
}
\description{
\code{plotWithManyBars} returns a barplot with the data provided to it.
}
\details{
Missing values in both \code{sortedData} and \code{sortedDataFrame} are
discarded (to avoid them being chosen as \code{maxSortedData}). Also, 
\code{highlightYear} is checked to ensure it is a valid year.

A barplot is created. Bars are created for all provided data. Unless
\code{showAllLabels} is set to \code{TRUE}, labels are only shown for the
highest and lowest values (and ties), as well as the \code{highlightYear}'s
data.
}
\examples{
# How do I show an example?
}

