% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotytdprecipitation.R
\name{plotYTDPrecipitation}
\alias{plotYTDPrecipitation}
\title{Plot year-to-date precipitation}
\usage{
plotYTDPrecipitation(plotMonth = format(orfwx::yesterdate(), "\%m"))
}
\arguments{
\item{plotMonth}{(optional) The month for which precipitation records are
desired. Defaults to the current month (except on the 1st of the month,
when it defaults to the previous month).}
}
\value{
Returns a plot.
}
\description{
\code{plotYTDPrecipitation} creates a plot of a month. with the precipitation
of the year to date.
}
\details{
Three lines are displayed: a red line with the highest year-to-date
precipitation recorded by that point in the month, a blue line with the
lowest year-to-date precipitation recorded by that point in the month, and
a black line with the curent year-to-date precipitation.
}
\examples{
\dontrun{
plotYTDPrecipitation()}
}

