% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removecsvprecipitation.R
\name{removeCsvPrecipitation}
\alias{removeCsvPrecipitation}
\title{Remove original precipitation variable from a data frame}
\usage{
removeCsvPrecipitation(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the \code{CsvPrecipitation}
variable is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeCsvPrecipitation} removes from a data frame the
\code{CsvPrecipitation} variable.
}
\details{
This function may be used to keep this information from cluttering up 
display, if the \code{CsvPrecipitation} is no longer needed.
}
\examples{
removeCsvPrecipitation(airportData)
}

