% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removecsvsnowfall.R
\name{removeCsvSnowfall}
\alias{removeCsvSnowfall}
\title{Remove extra date variables from a data frame}
\usage{
removeCsvSnowfall(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeCsvSnowfall} removes from a data frame the
\code{CsvSnowfall} variable.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{CsvSnowfall} is no longer needed.
}
\examples{
\dontrun{removeCsvSnowfall(bothStations)}
}

