% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removecsvvariables.R
\name{removeCsvVariables}
\alias{removeCsvVariables}
\title{Remove original precipitation and snowfall variables from a data frame}
\usage{
removeCsvVariables(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the variables are removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeCsvVariables} removes from a data frame the
\code{CsvPrecipitation} and \code{CsvSnowfall} variables.
}
\details{
This function may be used to keep them from cluttering up display, if the
those variables are no longer needed.
}
\examples{
\dontrun{removeCsvVariables(bothStations)}
}

