% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeextradatevariables.R
\name{removeExtraDateVariables}
\alias{removeExtraDateVariables}
\title{Remove extra date variables from a data frame}
\usage{
removeExtraDateVariables(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeExtraDateVariables} removes from a data frame the date
variables which have been added by \code{\link{computeExtraDateVariables}}.
}
\details{
These variables are integers for convenience use in filtering, and this
function may be used to keep them from cluttering up display.

Errors will result if any of the variables (\code{Year}, \code{Month},
\code{DayOfMonth}, and \code{DayOfYear}) have been renamed or deleted.
}
\examples{
removeExtraDateVariables(computeExtraDateVariables(airportData))
}

