% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeleapyearawaredayofyear.R
\name{removeLeapYearAwareDayOfYear}
\alias{removeLeapYearAwareDayOfYear}
\title{Remove extra date variables from a data frame}
\usage{
removeLeapYearAwareDayOfYear(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the 
\code{LeapYearAwareDayOfYear} variables is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeLeapYearAwareDayOfYear} removes from a data frame the
\code{LeapYearAwareDayOfYear} variable which was added by 
\code{\link{computeLeapYearAwareDayOfYear}}.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{LeapYearAwareDayOfYear} is no longer needed.
}
\examples{
\dontrun{
removeLeapYearAwareDayOfYear(computeLeapYearAwareDayOfYear(bothStations))
}
}

