% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeplotfiles.R
\name{removePlotFiles}
\alias{removePlotFiles}
\title{Deletes plot files from working directory}
\usage{
removePlotFiles(rpfDir = paste0(getwd(), "/"))
}
\arguments{
\item{rpfDir}{(optional) The directory to remove files from (defaults to the
working directory).}
}
\value{
Deletes files from working directory.
}
\description{
\code{removePlotFiles} deletes \code{*.png} files from the working directory.
}
\details{
This function deletes all PNG files from the working directory.
}
\examples{
\dontrun{
removePlotFiles()
}
}

