% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeprecipitationinches.R
\name{removePrecipitationInches}
\alias{removePrecipitationInches}
\title{Remove extra date variables from a data frame}
\usage{
removePrecipitationInches(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removePrecipitationInches} removes from a data frame the
\code{PrecipitationInches} variable.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{PrecipitationInches} is no longer needed.
}
\examples{
\dontrun{removePrecipitationInches(bothStations)}
}

