% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removesnowfallinches.R
\name{removeSnowfallInches}
\alias{removeSnowfallInches}
\title{Remove extra date variables from a data frame}
\usage{
removeSnowfallInches(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeSnowfallInches} removes from a data frame the
\code{SnowfallInches} variable.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{SnowfallInches} is no longer needed.
}
\examples{
\dontrun{removeSnowfallInches(bothStations)}
}

