% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removetemperaturespread.R
\name{removeTemperatureSpread}
\alias{removeTemperatureSpread}
\title{Remove extra date variables from a data frame}
\usage{
removeTemperatureSpread(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeTemperatureSpread} removes from a data frame the
\code{TemperatureSpread} variable which was added by 
\code{\link{computeTemperatureSpread}}.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{TemperatureSpread} is no longer needed.
}
\examples{
\dontrun{removeTemperatureSpread(computeTemperatureSpread(bothStations))}
}

