% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removewithprecipitation.R
\name{removeWithPrecipitation}
\alias{removeWithPrecipitation}
\title{Remove extra date variables from a data frame}
\usage{
removeWithPrecipitation(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeWithPrecipitation} removes from a data frame the
\code{WithPrecipitation} variable.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{WithPrecipitation} is no longer needed.
}
\examples{
\dontrun{removeWithPrecipitation(bothStations)}
}

