% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removewithsnowfall.R
\name{removeWithSnowfall}
\alias{removeWithSnowfall}
\title{Remove extra date variables from a data frame}
\usage{
removeWithSnowfall(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeWithSnowfall} removes from a data frame the
\code{WithSnowfall} variable.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{WithSnowfall} is no longer needed.
}
\examples{
\dontrun{removeWithSnowfall(bothStations)}
}

