% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singledaysweather.R
\name{singleDaysWeather}
\alias{singleDaysWeather}
\title{Create a data frame with a single day's weather}
\usage{
singleDaysWeather(highTemperature = NA, lowTemperature = NA,
  averageTemperature = NA, precipitation = NA, snowfall = NA,
  dataDate = Sys.Date())
}
\arguments{
\item{highTemperature}{An integer representing the day's high temperature
(in °F).}

\item{averageTemperature}{A number representing the day's average
temperature (in °F).}

\item{precipitation}{A number representing the day's precipitation (in 
inches).}

\item{snowfall}{A number representing the day's precipitation (in inches).}

\item{dataDate}{A date for the data represented (defaults to current day).}

\item{lowTemeprature}{An integer representing the day's low temperature (in
°F).}
}
\value{
Returns a data frame.
}
\description{
\code{singleDaysWeather} returns a data frame.
}
\details{
Details section to be written.
}
\examples{
singleDaysWeather(55, 34, 44.5, 0.00, 0.0, searchDate(12, 1, 2016))
}

