% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickfifth.R
\name{tickFifth}
\alias{tickFifth}
\title{Add ticks to y-axis of grid.}
\usage{
tickFifth()
}
\description{
\code{tickFifth} adds five smaller ticks in between the full ticks.
}
\details{
The smaller ticks are one-half the size of the full ticks.
}
\examples{
\dontrun{tickFifth()}
}

