% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickhalf.R
\name{tickHalf}
\alias{tickHalf}
\title{Add ticks to y-axis of grid.}
\usage{
tickHalf()
}
\description{
\code{tickHalf} adds a smaller tick in between the full ticks.
}
\details{
The smaller ticks are two-thirds the size of the full ticks.
}
\examples{
\dontrun{tickHalf()}
}

