% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticktenth.R
\name{tickTenth}
\alias{tickTenth}
\title{Add ticks to y-axis of grid.}
\usage{
tickTenth()
}
\description{
\code{tickTenth} adds ten smaller ticks in between the full ticks.
}
\details{
The smaller ticks are one-third the size of the full ticks.
}
\examples{
\dontrun{tickTenth()}
}
