% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxupdateclone.R
\name{wxUpdateClone}
\alias{wxUpdateClone}
\title{Retrieves a copy of the repository to upload updated weather data.}
\usage{
wxUpdateClone()
}
\value{
Downloads a repository to the \code{~/.orfwx/wxupdaterepo/} directory.
}
\description{
\code{wxUpdateClone} clones the Git repository with the \code{updates.csv} file
contained in it.
}
\details{
Details section to be written.
}
\examples{
\dontrun{
wxUpdateClone()
}
}
