% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesterdate.R
\name{yesterdate}
\alias{yesterdate}
\title{Yesterday's date}
\usage{
yesterdate()
}
\value{
Returns a date
}
\description{
\code{yesterdate} returns the previous day's date.
}
\details{
This is a convenience function since many times, functions are run for the
previous day's information.
}
\examples{
yesterdate()
}
