% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computecumulativepreciprecords.R
\name{computeCumulativePrecipRecords}
\alias{computeCumulativePrecipRecords}
\title{Compute cumulative precipitation records}
\usage{

  computeCumulativePrecipRecords(originalFrame = orfwx::computeCumulativePrecipitation(),
  ccprMonth = format(orfwx::yesterdate(), "\%m"), showYear = FALSE,
  includeNormals = FALSE,
  showLeapDay = orfwx::is.leapYear(as.integer(format(orfwx::yesterdate(),
  "\%Y"))), showLastCompleteMonth = FALSE)
}
\arguments{
\item{originalFrame}{(optional) The data frame to which the \code{MTDPrecip} 
and \code{YTDPrecip} variables are appended. Defaults to \code{allData()}.}

\item{ccprMonth}{(optional) The month for which precipitation records are
desired. Defaults to the current month (except on the 1st of the month,
when it defaults to the previous month).}

\item{showYear}{(optional) Whether or not to show the current or most recent
year's data for comparison (defaults to \code{FALSE}).}

\item{includeNormals}{(optional) Whether or not to show the 1981-2010 
normals for comparison (defaults to \code{FALSE}).}

\item{showLeapDay}{(optional) Whether or not to show data for February 29th
(defaults to \code{FALSE}, unless the current year is a leap year).}

\item{showLastCompleteMonth}{(optional) Whether to show data from the last
complete month or whether to show the month in progress (defaults to
\code{FALSE}).}
}
\value{
Returns a data frame.
}
\description{
\code{computeCumulativePrecipRecords} returns a data frame with
the maximum and minimum values of \code{MTDPrecip}.
}
\details{
This function probably contains too much for one function.

The function:

\enumerate{
  \item Checks to make sure that \code{ccprMonth} is an integer,
  \item Sets the year to display (if \code{showYear} is \code{TRUE})
  \item Changes \code{originalFrame} to add date variables and include only
    data from the same calendar month.
  \item Creates \code{recordsFrame} (a copy of \code{originalFrame}), and 
    uses it to determine the records for month-to-date and year-to-date 
    (excluding the plotted year from the records).
  \item Creates \code{yearsFrame} with a row for each day of the month, and
    initializes variables.
  \item Loops through \code{yearsFrame} to get the most recent year for each
    day's record
}
}
\examples{
\dontrun{computeCumulativePrecipRecords()}
}

