% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayeachyear.R
\name{dayEachYear}
\alias{dayEachYear}
\title{Find the weather for a give day each year.}
\usage{
dayEachYear(wxUniverse = orfwx::allData(), deyMonth, deyDayOfMonth)
}
\arguments{
\item{wxUniverse}{The data frame with the weather information we will be
taking the data from.}

\item{deyMonth}{An integer representing the month of the year (1=January,
12=December) for which the date we are interested in.}

\item{deyDayOfMonth}{An integer representing the day of the month (1-31) for
the date we are interested in.}
}
\value{
A data frame is returned.
}
\description{
\code{dayEachYear} takes a data frame with weather information and returns only
the weather for one particular day of each year.
}
\details{
Details section to be written.
}
\examples{
dayEachYear(deyMonth = 2, deyDayOfMonth = 15)  # Weather for February 15th
}
