% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getupdatedcsv.R
\name{getUpdatedCsv}
\alias{getUpdatedCsv}
\title{Retrieves a file of updated weather observations}
\usage{
getUpdatedCsv()
}
\value{
Downloads a file to the \code{~/.orfwx} directory.
}
\description{
\code{getUpdatedCsv} downloads a CSV file with weather observations.
}
\details{
Most observations are included within the `orfwx` package. (See 
\code{\link{airportData}}, \code{\link{bothStations}}, and 
\code{\link{earlyDowntownData}}).  This function is used to retrieve data
newer than that included within the package.

\code{getUpdatedCsv} uses \code{httr} to download an
\code{updates.csv} file. It is stored in the \code{~/.orfwx} directory,
which this function attempts to create if it is not already present.
}
\examples{
\dontrun{
getUpdatedCsv()
}
}

