% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalsuffix.R
\name{ordinalSuffix}
\alias{ordinalSuffix}
\title{Add an ordinal suffix to an integer}
\usage{
ordinalSuffix(osInteger)
}
\arguments{
\item{osInteger}{The number to convert.}
}
\value{
Returns a string.
}
\description{
\code{ordinalSuffix} returns the integer provided to it as a string, with the
ordinal suffix (\emph{st}, \emph{nd}, \emph{rd}, or \emph{th}) appended.
}
\details{
This function's workflow:
\enumerate{
\item Is \code{osInteger} a positive integer? If not, a warning is issued and this
function returns \code{osInteger} unchanged.
\item Determine what the proper suffix should be:
\enumerate{
\item If the \emph{tens} digit is a 1 (e.g., 10 through 19, 110 through 19, etc.),
the suffix is \code{th}.
\item In all other cases, look to the result of \code{osInteger} mod 10:
\itemize{
\item If 1, suffix is \code{st},
\item If 2, suffix is `nd``,
\item If 3, suffix is `rd``, and
\item In all other cases, suffix is \code{th}.
}
}
\item Paste together \code{osInteger} with the suffix and return this as a string.
}
}
\examples{
orfwx:::ordinalSuffix(1)  # Returns "1st"
orfwx:::ordinalSuffix(2)  # Returns "2nd"
orfwx:::ordinalSuffix(3)  # Returns "3rd"
orfwx:::ordinalSuffix(4)  # Returns "4th"
orfwx:::ordinalSuffix(11)  # Returns "11th"
}
