% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padsingledigitinteger.R
\name{padSingleDigitInteger}
\alias{padSingleDigitInteger}
\title{Pad a number with a zero to create a two-character string (if single digit).}
\usage{
padSingleDigitInteger(theInteger)
}
\arguments{
\item{theInteger}{Any integer between 0 and 99 (inclusive).}
}
\value{
Returns a two-character string.
}
\description{
\code{padSingleDigitInteger} returns a two-character string, padding a
single-digit integer with a leading zero, if required.
}
\details{
This function converts \code{theInteger} into an integer and
returns that integer if between 10 and 99 (inclusive); otherwise, 
it adds a leading zero, to create a two-character string.

Errors will be provided for negative integers and integers with
more than two digits.
}
\examples{
padSingleDigitInteger(20)  # "20"
padSingleDigitInteger(9)  # "09"
padSingleDigitInteger(0)  # "00"
}

