% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoolestmaxtempoverhistory.R
\name{plotCoolestMaxTempOverHistory}
\alias{plotCoolestMaxTempOverHistory}
\title{Plot coolest historical high tempertures on a single date}
\usage{
plotCoolestMaxTempOverHistory(plotDate = searchDate(), daysWeather = NULL)
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}

\item{daysWeather}{(optional) The weather for a date not yet included in
the \code{mutatedBothStations} dataset, usually passed by the
\code{singleDaysWeather} function.}
}
\value{
Returns a barplot.
}
\description{
\code{plotCoolestMaxTempOverHistory} returns a barplot with the coolest 
high temperatures of one day of the year.
}
\details{
This code makes my head hurt, because it was some of the earliest R code I
wrote, when I was mostly in make-it-work mode, with huge gaps in my
understanding of R (and of barplot particularly)
}
\examples{
plotCoolestMaxTempOverHistory(searchDate(11, 26))  # plot for November 26th
}

