% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoolestmintempoverhistory.R
\name{plotCoolestMinTempOverHistory}
\alias{plotCoolestMinTempOverHistory}
\title{Plot coolest historical low tempertures on a single date}
\usage{
plotCoolestMinTempOverHistory(plotDate = searchDate(), daysWeather = NULL)
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}

\item{daysWeather}{(optional) The weather for a date not yet included in
the \code{mutatedBothStations} dataset, usually passed by the
\code{singleDaysWeather} function.}
}
\value{
Returns a barplot.
}
\description{
\code{plotCoolestMinTempOverHistory} returns a barplot with the coolest 
low temperatures of one day of the year.
}
\details{
If \code{daysWeather} is passed (usually using 
\code{\link{singleDaysWeather}}), the function ensures that its weather is 
for the same date as that provided by \code{plotDate} (either passed to the 
function or the default value of the current system date). If they do not 
match, the function terminates with an error message.

The data from \code{daysWeather} is added to that provided by 
\code{\link{mutatedBothStations}} and the weather on the same date each year
is compared.

A barplot is plotted by \code{\link{plotWithManyBars}} with the low 
temperatures for all days with a low temperature less than or equal to that 
provided by \code{daysWeather}. (In all cases, the ten coolest years are 
shown, along with any ties, with additional years added until the year 
provided by \code{daysWeather} is shown, and also all years tied with the 
same temperature.) The bar for the year provided by \code{daysWeather} (or,
if \code{daysWeather} is not provided, the current year) is highlighted with
a bar of a different color in the barplot.
}
\examples{
plotCoolestMaxTempOverHistory(searchDate(11, 26))  # plot for November 26th
}

