% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoolestmintempoverhistory.R
\name{plotCoolestMinTempOverHistory}
\alias{plotCoolestMinTempOverHistory}
\title{Plot coolest historical low tempertures on a single date}
\usage{
plotCoolestMinTempOverHistory(plotDate = searchDate())
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}
}
\value{
Returns a barplot.
}
\description{
\code{plotCoolestMinTempOverHistory} returns a barplot with the coolest 
loq temperatures of one day of the year.
}
\details{
This code makes my head hurt, because it was some of the earliest R code I
wrote, when I was mostly in make-it-work mode, with huge gaps in my
understanding of R (and of barplot particularly)
}
\examples{
plotCoolestMaxTempOverHistory(searchDate(11, 26))  # plot for November 26th
}

