% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmintempoverhistory.R
\name{plotMinTempOverHistory}
\alias{plotMinTempOverHistory}
\title{Plot historical low tempertures}
\usage{
plotMinTempOverHistory(plotDate = searchDate(), daysWeather = NULL)
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}

\item{daysWeather}{(optional) The weather for a date not yet included in
the \code{mutatedBothStations} dataset, usually passed by the
\code{singleDaysWeather} function.}
}
\value{
Returns a barplot.
}
\description{
\code{plotMinTempOverHistory} returns a barplot with the low temperatures
of one day each year.
}
\details{
This code makes my head hurt, because it was some of the earliest R code I
wrote, when I was mostly in make-it-work mode, with huge gaps in my
understanding of R (and of barplot particularly)
}
\examples{
plotMinTempOverHistory(searchDate(11, 26))  # plot for November 26th
}

