% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmtdprecipitation.R
\name{plotMTDPrecipitation}
\alias{plotMTDPrecipitation}
\title{Plot month-to-date precipitation}
\usage{
plotMTDPrecipitation(plotMonth = format(orfwx::yesterdate(), "\%m"),
  saveToFile = FALSE)
}
\arguments{
\item{plotMonth}{(optional) The month for which precipitation records are
desired. Defaults to the current month (except on the 1st of the month,
when it defaults to the previous month).}

\item{saveToFile}{(optional) Writes plot to a PNG file (defaults to 
\code{FALSE}).}
}
\value{
Returns a plot.
}
\description{
\code{plotMTDPrecipitation} creates a plot with the precipitation
of the month to date.
}
\details{
Three lines are displayed: a red line with the highest month-to-date
precipitation recorded by that point in the month, a blue line with the
lowest month-to-date precipitation recorded by that point in the month, and
a black line with the curent year's month-to-date precipitation.
}
\examples{
\dontrun{
plotMTDPrecipitation()}
}

