% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwarmestmintempoverhistory.R
\name{plotWarmestMinTempOverHistory}
\alias{plotWarmestMinTempOverHistory}
\title{Plot warmest historical low tempertures on a single date}
\usage{
plotWarmestMinTempOverHistory(plotDate = searchDate())
}
\arguments{
\item{plotDate}{(optional) The date to be searched for, defaulting to the 
current date.}
}
\value{
Returns a barplot.
}
\description{
\code{plotWarmestMinTempOverHistory} returns a barplot with the warmest 
low temperatures of one day of the year.
}
\details{
This code makes my head hurt, because it was some of the earliest R code I
wrote, when I was mostly in make-it-work mode, with huge gaps in my
understanding of R (and of barplot particularly)
}
\examples{
plotWarmestMinTempOverHistory(searchDate(11, 27))  # plot for November 27th
}

