% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwithmanybars.R
\name{plotWithManyBars}
\alias{plotWithManyBars}
\title{Create a barplot with temperature or precipitation data}
\usage{
plotWithManyBars(sortedData, sortedDataFrame, titlePlotWmb, yAxisLabelPlotWmb,
  plottingPrecip = FALSE, showAllLabels = FALSE)
}
\arguments{
\item{sortedData}{A vector used for the values to be shown on the barplot.}

\item{sortedDataFrame}{A data frame, from which \code{sortedData} has
usually been excerpted.}

\item{titlePlotWmb}{The title of the barplot}

\item{yAxisLabelPlotWmb}{The label for the y axis}

\item{plottingPrecip}{A logical value indicating if we are dealing with
precipitation}

\item{showAllLabels}{A logical value indicating if we are showing all
labels (if \code{FALSE}, labels will only be shown on the max and min
values and on the current year)}
}
\value{
Returns a barplot.
}
\description{
\code{plotWithManyBars} returns a barplot with the data provided to it.
}
\details{
This code makes my head hurt, because it was some of the earliest R code I
wrote, when I was mostly in make-it-work mode, with huge gaps in my
understanding of R (and of barplot particularly)
}
\examples{
# How do I show an example?
}

