% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removecsvprecipitation.R
\name{removeCsvPrecipitation}
\alias{removeCsvPrecipitation}
\title{Remove extra date variables from a data frame}
\usage{
removeCsvPrecipitation(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the date variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeCsvPrecipitation} removes from a data frame the
\code{CsvPrecipitation} variable.
}
\details{
This function may be used to keep them from cluttering up display, if the
\code{CsvPrecipitation} is no longer needed.
}
\examples{
\dontrun{removeCsvPrecipitation(bothStations)}
}

