% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removecsvsnowfall.R
\name{removeCsvSnowfall}
\alias{removeCsvSnowfall}
\title{Remove original snowfall variable from a data frame}
\usage{
removeCsvSnowfall(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the \code{CsvSnowfall}
variable is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeCsvSnowfall} removes from a data frame the
\code{CsvSnowfall} variable.
}
\details{
This function may be used to keep this information from cluttering up 
display, if the \code{CsvSnowfall} is no longer needed.
}
\examples{
removeCsvSnowfall(bothStations)
}

