% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeinchesvariables.R
\name{removeInchesVariables}
\alias{removeInchesVariables}
\title{Remove inches variables from a data frame}
\usage{
removeInchesVariables(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the inches variables are 
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeInchesVariables} removes from a data frame the
\code{PrecipitationInches} and  \code{SnowfallInches} variables.
}
\details{
This function may be used to keep these variables from cluttering up
display, if they are no longer needed.
}
\examples{
removeInchesVariables(convertCsvToNumericAndLogical(airportData))
}

