% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeprecipitationinches.R
\name{removePrecipitationInches}
\alias{removePrecipitationInches}
\title{Remove precipitation amount variable from a data frame}
\usage{
removePrecipitationInches(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the
\code{PrecipitationInches} variable is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removePrecipitationInches} removes from a data frame the
\code{PrecipitationInches} variable.
}
\details{
This function may be used to keep the \code{PrecipitationInches} variable
from cluttering up display, if it is no longer needed.
}
\examples{
removePrecipitationInches(convertCsvToNumericAndLogical(airportData))
}

