% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removesnowfallinches.R
\name{removeSnowfallInches}
\alias{removeSnowfallInches}
\title{Remove snowfall amount variable from a data frame}
\usage{
removeSnowfallInches(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the \code{SnowfallInches}
variable is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeSnowfallInches} removes from a data frame the
\code{SnowfallInches} variable.
}
\details{
This function may be used to keep the \code{SnowfallInches} variable from
cluttering up display, if it is no longer needed.
}
\examples{
removeSnowfallInches(convertCsvToNumericAndLogical(airportData))
}

