% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removetemperaturespread.R
\name{removeTemperatureSpread}
\alias{removeTemperatureSpread}
\title{Remove temperature spread variable from a data frame}
\usage{
removeTemperatureSpread(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the \code{TemperatureSpread} 
variable is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeTemperatureSpread} removes from a data frame the
\code{TemperatureSpread} variable which was added by 
\code{\link{computeTemperatureSpread}}.
}
\details{
This function may be used to keep the \code{TemperatureSpread} variable
from cluttering up display, if it is no longer needed.
}
\examples{
removeTemperatureSpread(computeTemperatureSpread(airportData))
}

