% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removewithprecipitation.R
\name{removeWithPrecipitation}
\alias{removeWithPrecipitation}
\title{Remove logical precipitation variable from a data frame}
\usage{
removeWithPrecipitation(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the \code{WithPrecipitation}
variable is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeWithPrecipitation} removes from a data frame the
\code{WithPrecipitation} variable.
}
\details{
This function may be used to keep the \code{WithPrecipitation} variable
from cluttering up display, if it is no longer needed.
}
\examples{
removeWithPrecipitation(convertCsvToNumericAndLogical(airportData))
}
