% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removewithsnowfall.R
\name{removeWithSnowfall}
\alias{removeWithSnowfall}
\title{Remove logical snowfall variable from a data frame}
\usage{
removeWithSnowfall(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the \code{WithSnowfall} 
variable is removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeWithSnowfall} removes from a data frame the
\code{WithSnowfall} variable.
}
\details{
This function may be used to keep the \code{WithSnowfall} variable from
cluttering up display, if it is no longer needed.
}
\examples{
removeWithSnowfall(convertCsvToNumericAndLogical(airportData))
}

