% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removewithvariables.R
\name{removeWithVariables}
\alias{removeWithVariables}
\title{Remove logical variables from a data frame}
\usage{
removeWithVariables(originalFrame)
}
\arguments{
\item{originalFrame}{The data frame from which the logical variables are
removed.}
}
\value{
Returns a data frame.
}
\description{
\code{removeWithVariables} removes from a data frame the
\code{WithPrecipitation} and \code{WithSnowfall} variables.
}
\details{
This function may be used to keep them from cluttering up display, if these
variables are no longer needed.
}
\examples{
removeWithVariables(convertCsvToNumericAndLogical(airportData))
}
